Large change 1

+++ SamSoarII.Simulation/Core/VariableModel/SimulateDWordModel.cs
@@ -9,11 +9,11 @@ namespace SamSoarII.Simulation.Core.VariableModel
 {
     public class SimulateDWordUnit : SimulateVariableUnit
     {
-        protected int value;
+        protected long value;
         override public Object Value
         {
             get { return (Object)(this.value); }
-            set { this.value = (int)(value); }
+            set { this.value = (long)(value); }
         }
         public override string ToString()
         {
@@ -63,8 +63,8 @@ namespace SamSoarII.Simulation.Core.VariableModel
 
         public override void Update(SimulateDllModel dllmodel)
         {
-            int[] ivalues = dllmodel.GetValue_DWord(Name, 1);
-            int value_old = value;
+            long[] ivalues = dllmodel.GetValue_DWord(Name, 1);
+            long value_old = value;
             this.value = ivalues[0];
             if (value_old != value)
             {
@@ -73,7 +73,7 @@ namespace SamSoarII.Simulation.Core.VariableModel
         }
         public override void Set(SimulateDllModel dllmodel)
         {
-            int[] ivalues = { this.value };
+            long[] ivalues = { this.value };
             dllmodel.SetValue_DWord(Name, 1, ivalues);
         }
     }
@@ -114,7 +114,7 @@ namespace SamSoarII.Simulation.Core.VariableModel
 
         public override void Update(SimulateDllModel dllmodel)
         {
-            int[] ivalues = dllmodel.GetValue_DWord(Name, Size);
+            long[] ivalues = dllmodel.GetValue_DWord(Name, Size);
             for (int i = 0; i < size; i++)
             {
                 this.values[i].Value = ivalues[i];
@@ -123,7 +123,7 @@ namespace SamSoarII.Simulation.Core.VariableModel
 
         public override void Set(SimulateDllModel dllmodel)
         {
-            int[] ivalues = new int[size];
+            long[] ivalues = new long[size];
             for (int i = 0; i < size; i++)
             {
                 ivalues[i] = (int)(this.values[i].Value);
